+++ include/engrave.h
@@ -10,7 +10,7 @@ struct engr {
 	char *engr_txt;
 	xchar engr_x, engr_y;
 	unsigned engr_lth;	/* for save & restore; not length of text */
-	int32_t engr_time;		/* moment engraving was (will be) finished */
+	long engr_time;		/* moment engraving was (will be) finished */
 	xchar engr_type;
 #define DUST	   1
 #define ENGRAVE    2
